\name{data_bbsalaries}
\alias{data_bbsalaries}
\docType{data}
\title{Baseball pitcher salary data}
\description{
Salaries of 176 piters for the 1987 season. 
The variables are as follows:
}
\usage{data(data_bbsalaries)}
\format{
  A data frame with 176 observations on the following 2 variables.
  \describe{
    \item{\code{y}}{Log of the base salary in dollars}
    \item{\code{x}}{Log of the number of years experience}
  }
}

\source{
Hettmansperger, T.P. and McKean J.W. (2011), Robust Nonparametric Statistical Methods, 2nd ed., New York: Chapman-Hall.
}
\references{
Hettmansperger, T.P. and McKean J.W. (2011), Robust Nonparametric Statistical Methods, 2nd ed., New York: Chapman-Hall.
}
\examples{
data(data_bbsalaries)
## maybe str(data_bbsalaries) ; plot(data_bbsalaries) ...
}
\keyword{datasets}
